telemedicinebackend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
MarkupSafe==3.0.3
multidict==6.7.0
mysqlclient==2.2.7
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-http-client==3.3.7
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
sendgrid==6.12.5
sqlparse==0.5.3
twilio==9.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
Werkzeug==3.1.4
yarl==1.22.0
